+++ NEWS.md
@@ -5,6 +5,8 @@
 * BUG fix in 'pivotatStat' (cont'd): the pivotal statistic was not
   calculated properly for kFWER threshold family when the cardinal of
   the set of candidate hypotheses ($\cC$) was less than m.
+* Added a test script to check that the R and C++ versions of
+  'minPseudoRanks' give identical results.
 
 ## Version 0.4.3 [2016-06-09]